+++ scripts/download-assets.js
@@ -4,20 +4,20 @@ const fs = require("fs");
 const NPM_URL = "https://registry.npmjs.org";
 const ROOT = "public/assets";
 
-const FFMPEG_VERSION = "0.13.0";
-const UTIL_VERSION = "0.13.0";
-const CORE_VERSION = "0.13.0";
-const CORE_MT_VERSION = "0.13.0";
+const FFMPEG_VERSION = "0.13.1";
+const UTIL_VERSION = "0.13.1";
+const CORE_VERSION = "0.13.1";
+const CORE_MT_VERSION = "0.13.1";
 
-const FFMPEG_TGZ = `ffmpeg-${FFMPEG_VERSION}.tgz`;
-const UTIL_TGZ = `util-${UTIL_VERSION}.tgz`;
-const CORE_TGZ = `core-${CORE_VERSION}.tgz`;
-const CORE_MT_TGZ = `core-mt-${CORE_MT_VERSION}.tgz`;
+const FFMPEG_TGZ = `ffmpeg-wasm-${FFMPEG_VERSION}.tgz`;
+const UTIL_TGZ = `ffmpeg-wasm-util-${UTIL_VERSION}.tgz`;
+const CORE_TGZ = `ffmpeg-wasm-core-${CORE_VERSION}.tgz`;
+const CORE_MT_TGZ = `ffmpeg-wasm-core-mt-${CORE_MT_VERSION}.tgz`;
 
-const FFMPEG_TGZ_URL = `${NPM_URL}/@project516/ffmpeg/-/${FFMPEG_TGZ}`;
-const UTIL_TGZ_URL = `${NPM_URL}/@project516/util/-/${UTIL_TGZ}`;
-const CORE_TGZ_URL = `${NPM_URL}/@project516/core/-/${CORE_TGZ}`;
-const CORE_MT_TGZ_URL = `${NPM_URL}/@project516/core-mt/-/${CORE_MT_TGZ}`;
+const FFMPEG_TGZ_URL = `${NPM_URL}/@project516/ffmpeg-wasm/-/${FFMPEG_TGZ}`;
+const UTIL_TGZ_URL = `${NPM_URL}/@project516/ffmpeg-wasm-util/-/${UTIL_TGZ}`;
+const CORE_TGZ_URL = `${NPM_URL}/@project516/ffmpeg-wasm-core/-/${CORE_TGZ}`;
+const CORE_MT_TGZ_URL = `${NPM_URL}/@project516/ffmpeg-wasm-core-mt/-/${CORE_MT_TGZ}`;
 
 const mkdir = (dir) => {
   !fs.existsSync(dir) && fs.mkdirSync(dir);
@@ -26,7 +26,7 @@ const mkdir = (dir) => {
 const downloadAndUntar = async (url, tgzName, dst) => {
   const dir = `${ROOT}/${dst}`;
   if (fs.existsSync(dir)) {
-    console.log(`found @project516/${dst} assets.`);
+    console.log(`found ${dst} assets.`);
     return;
   }
   console.log(`download and untar ${url}`);